rewrite DriveItem hierarchy, fix a error related with ByteBufStream
Rewrite DriveItem hierarchy
 Now support complex item type as RemoteFileItem and RemoteFolderItem

+++ src/main/java/com/bhyoo/onedrive/container/facet/AudioFacet.java
@@ -17,17 +17,17 @@ public class AudioFacet {
 	@Getter @Setter(PRIVATE) @Nullable protected String album;
 	@Getter @Setter(PRIVATE) @Nullable protected String albumArtist;
 	@Getter @Setter(PRIVATE) @Nullable protected String artist;
-	@Getter @Setter(PRIVATE) @Nullable protected Long bitrate;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer bitrate;
 	@Getter @Setter(PRIVATE) @Nullable protected String composers;
 	@Getter @Setter(PRIVATE) @Nullable protected String copyright;
-	@Getter @Setter(PRIVATE) @Nullable protected Long disc;
-	@Getter @Setter(PRIVATE) @Nullable protected Long discCount;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer disc;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer discCount;
 	@Getter @Setter(PRIVATE) @Nullable protected Long duration;
 	@Getter @Setter(PRIVATE) @Nullable protected String genre;
 	@Getter @Setter(PRIVATE) @Nullable protected Boolean hasDrm;
 	@Getter @Setter(PRIVATE) @Nullable protected Boolean isVariableBitrate;
 	@Getter @Setter(PRIVATE) @Nullable protected String title;
-	@Getter @Setter(PRIVATE) @Nullable protected Long track;
-	@Getter @Setter(PRIVATE) @Nullable protected Long trackCount;
-	@Getter @Setter(PRIVATE) @Nullable protected Long year;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer track;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer trackCount;
+	@Getter @Setter(PRIVATE) @Nullable protected Integer year;
 }
